sonar fixes
Very interesting was the exception swallowing in PackagePropertiesPart. When it was properly thrown, it already led to various errors in the junits test - I've fixed the handling for at least the ones which are in our test set

git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1706169 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/poifs/filesystem/BlockStore.java
@@ -80,7 +80,7 @@ public abstract class BlockStore {
     protected class ChainLoopDetector {
        private boolean[] used_blocks;
        protected ChainLoopDetector(long rawSize) {
-          int numBlocks = (int)Math.ceil( rawSize / getBlockStoreBlockSize() );
+          int numBlocks = (int)Math.ceil( ((double)rawSize) / getBlockStoreBlockSize() );
           used_blocks = new boolean[numBlocks];
        }
        protected void claim(int offset) {
