+++ src/java/org/apache/poi/poifs/property/NPropertyTable.java
@@ -120,7 +120,7 @@ public final class NPropertyTable extends PropertyTableBase {
     public int countBlocks()
     {
        int size = _properties.size() * POIFSConstants.PROPERTY_SIZE;
-       return (int)Math.ceil(size / _bigBigBlockSize.getBigBlockSize());
+       return (int)Math.ceil( ((double)size) / _bigBigBlockSize.getBigBlockSize());
     }
  
     /**
